dockingpp/pyproject.toml
[project]
name = "dockingpp"
version = "0.1.0"
description = "Docking pipeline skeleton."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pydantic",
  "pyyaml",
  "tqdm",
]

[project.optional-dependencies]
scipy = ["scipy"]

[project.scripts]
dockingpp = "dockingpp.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
